fix: gfx ext buffer offset parameter 32 to 64 bit

+++ extensions/pl_graphics_cpu.c
@@ -880,7 +880,7 @@ pl_request_command_buffer(plCommandPool* ptPool, const char* pcDebugName)
 }
 
 void
-pl_copy_buffer(plBlitEncoder* ptEncoder, plBufferHandle tSource, plBufferHandle tDestination, uint32_t uSourceOffset, uint32_t uDestinationOffset, size_t szSize)
+pl_copy_buffer(plBlitEncoder* ptEncoder, plBufferHandle tSource, plBufferHandle tDestination, uint64_t uSourceOffset, uint64_t uDestinationOffset, size_t szSize)
 {
 }
